Canvas.pm: improve gocr output, less special casing for it
Integrate the logic for displaying gocr output on the ORC tab
in the the standard logic for displaying for OCR output.

While at it, improve readability of gocr output by increasing it
to a sensible level.

Finally, do not make the font used depending on the size or
orientation of the bounding box: scaling & rotating is done later.